% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Slice Rows from Dataframe Based on Variable Values}
\usage{
filter(data, ..., mode = any)
}
\arguments{
\item{data}{a dataframe to be assigned a groups attribute}

\item{...}{a series of \code{name == value} logical expressions. see examples below.}

\item{mode}{\code{any} or \code{all} to determine the selection mode.}
}
\value{
A dataframe subset of \code{data} that contains only those rows abiding by the selection mode and the logical expressions provided.
}
\description{
\code{filter} takes \code{data}, a dataframe object, and a series of \code{name == value} logical expressions. If \code{mode} is set to \code{any}, \code{filter} will produce a dataframe for which every row abides by at least one of the logical statements. If \code{mode} is set to \code{all}, \code{filter} will produce a dataframe for which every row abides by all of the logical statements.
}
\examples{
# any mode:
mtcars |> filter(gear==4, disp>385, mode=any)
# notice we get a data set containing all rows for which gear = 4 OR disp > 385

# all mode:
mtcars |> filter(gear>3, cyl==8, mode=all)
# notice we get a data set containing all rows for which gear > 3 AND cyl = 8
}
