% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group.by}
\alias{group.by}
\title{Group Dataframe Rows by Unique Values}
\usage{
group.by(data, col)
}
\arguments{
\item{data}{a dataframe to be assigned a groups attribute}

\item{col}{an unquoted expression for the name of a column in the dataframe}
}
\value{
The \code{data} dataframe, with a new \code{groups} attribute that can be used manually, but will primarily be used by other sidewindr functions.
}
\description{
\code{group.by} takes \code{data}, a dataframe object, and one column name in the dataframe. It will assign a \code{groups} attribute to the object. The \code{groups} attribute will be a list of double lists, where each list will contain all row indicies of one unique value of \code{col}.

Once the \code{groups} attribute has been set on a dataframe, passing it into other \code{sidewindr} functions will make them behave groupwise instead of against the whole dataset. \code{group.by} only supports single level grouping. Grouping an already grouped dataframe will overwrite the current groupings.
}
\examples{
mtcars |> group.by(drat)
# notice: group.by does not inherently do anything on its own!

# to visualize the effects, let's use slice.head,
mtcars |> group.by(cyl) |> slice.head(3)
# now we see 3 cars from each of the groups in mtcars!
}
