% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Create New Columns in a Dataframe}
\usage{
mutate(data, ...)
}
\arguments{
\item{data}{a dataframe to be assigned a groups attribute}

\item{...}{a series of \code{name = value} expressions. see examples below.}
}
\value{
A dataframe that contains all data from \code{data}, but has been augmented to contain those columns specified in \code{...} with values accordingly.
}
\description{
\code{mutate} takes \code{data}, a dataframe object, and a series of \code{name = value} pairs. \code{mutate} will then attempt to evaluate each of the provided \code{value} expressions and make it a new column in \code{data} under \code{name}. \code{mutate} makes no attempt to check for correct typing or defined operations when performing calculations, please ensure data validity before using.
}
\examples{
mtcars |> mutate(zoomies=qsec/hp)
# make a new variable "zoomies" equal to the quarter second time divided by the horsepower

mtcars |> mutate(egn.shape=ifelse(vs==0, "V-shaped","Straight"), transmission=ifelse(am==0,"Automatic","Manual"))
# convert the boolean vs and am values to strings using mutate and ifelse
}
