% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename.cols}
\alias{rename.cols}
\title{Rename Variables in Columns}
\usage{
rename.cols(data, ...)
}
\arguments{
\item{data}{a dataframe with columns to be renamed}

\item{...}{a series of \code{name = newname} expression pairs, where \code{name} is a column in \code{data}, and \code{newname} is the desired name of \code{name}. Neither \code{name} nor \code{newname} should be strings.}
}
\value{
The \code{data} dataframe with the specified \code{name} columns renamed to \code{newname}.
}
\description{
\code{rename.cols} takes \code{data}, a dataframe object, and \code{name = newname} expressions. \code{rename.cols} will try to reassign every \code{name} in \code{colnames(data)} to \code{newname}.
}
\examples{
mtcars |> rename.cols(mpg=Miles.Per.Gallon, hp=Horses.Per.Car)
# give the dataset some.. more clear names.
}
