% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_head.R
\name{slice.head}
\alias{slice.head}
\title{Pull Top Rows of a Dataframe}
\usage{
slice.head(data, n)
}
\arguments{
\item{data}{a dataframe from which rows should be sliced.}

\item{n}{a double representing the number of rows to slice.}
}
\value{
The \code{data} dataframe, now sliced to only contain \code{n} rows per group in the dataframe.

If \code{n} is larger than the size of the total dataframe, it will return the entire dataframe.

If \code{n} is larger than a group when performing the operation groupwise, it will select the entire group.
}
\description{
\code{slice.head} takes two parameters, a dataframe \code{data} and a number \code{n}. It will attempt to slice \code{n} rows out of the dataframe from the top of the dataframe and return these sliced rows as a new dataframe. If \code{data} is grouped, it will slice \code{n} rows for each group in \code{data}.

Note: Slicing a dataframe will remove the grouping indicies imposed by \code{group.by}. Please regroup dataframes following this operation.
}
\examples{
mtcars |> slice.head(5)
# slice the first 5 rows of mtcars

mtcars |> group.by(cyl) |> slice.head(2)
# slice the first 2 rows for each amount of cyl in the dataset

mtcars |> group.by(cyl) |> slice.head(15)
# notice: 15 is larger than the size of our groups, so some group slices may include the entire group!
}
